fix: when duplicating a metric, now the change is persisted.
