+++ frontend/src/scenes/experiments/MetricsView/DeltaChart.tsx
@@ -4,6 +4,8 @@ import { useActions, useValues } from 'kea'
 import { FEATURE_FLAGS } from 'lib/constants'
 import { createContext, useContext, useState } from 'react'
 
+import type { ExperimentMetric } from '~/queries/schema/schema-general'
+import { NodeKind } from '~/queries/schema/schema-general'
 import { Experiment, ExperimentIdType, FunnelExperimentVariant, InsightType, TrendExperimentVariant } from '~/types'
 
 import { experimentLogic } from '../experimentLogic'
@@ -16,6 +18,7 @@ import { GridLines } from './GridLines'
 import { MetricHeader } from './MetricHeader'
 import { MetricsChartLayout } from './MetricsChartLayout'
 import { SignificanceHighlight } from './SignificanceHighlight'
+import { getDefaultMetricTitle } from './utils'
 import { VariantTooltip } from './VariantTooltip'
 import { generateViolinPath } from './violinUtils'
 
@@ -398,6 +401,30 @@ function DeltaChartContent({ chartSvgRef }: { chartSvgRef: React.RefObject<SVGSV
     )
 }
 
+/**
+ * High order function that takes an experiment and the primary status of the metric, and returns
+ * a function that, given a metric, returns the updated slice of the experiment object.
+ */
+const duplicateMetric =
+    (experiment: Experiment, isPrimary: boolean) =>
+    (metric: ExperimentMetric): Partial<Experiment> => {
+        const name = metric.name
+            ? `${metric.name} (copy)`
+            : metric.kind === NodeKind.ExperimentMetric
+            ? `${getDefaultMetricTitle(metric)} (copy)`
+            : undefined
+
+        const newMetric = { ...metric, id: undefined, name }
+        const metricsKey = isPrimary ? 'metrics' : 'metrics_secondary'
+        const metrics = [...experiment[metricsKey]]
+        const originalIndex = metrics.findIndex((m) => m === metric)
+        metrics.splice(originalIndex + 1, 0, newMetric)
+
+        return {
+            [metricsKey]: metrics,
+        }
+    }
+
 // Main DeltaChart component
 export function DeltaChart({
     isSecondary,
@@ -437,7 +464,7 @@ export function DeltaChart({
         hasMinimumExposureForResults,
     } = useValues(experimentLogic)
 
-    const { openVariantDeltaTimeseriesModal } = useActions(experimentLogic)
+    const { openVariantDeltaTimeseriesModal, setExperiment, updateExperimentMetrics } = useActions(experimentLogic)
 
     // Loading state
     const resultsLoading = isSecondary ? secondaryMetricResultsLoading : metricResultsLoading
@@ -464,6 +491,8 @@ export function DeltaChart({
         return HORIZONTAL_PADDING + percentage * (VIEW_BOX_WIDTH - 2 * HORIZONTAL_PADDING)
     }
 
+    const onDuplicateMetric = duplicateMetric(experiment, !isSecondary)
+
     // Metric title panel
     const metricTitlePanel = (
         <MetricHeader
@@ -471,6 +500,10 @@ export function DeltaChart({
             metric={metric}
             metricType={metricType}
             isPrimaryMetric={!isSecondary}
+            onDuplicateMetricClick={(metric) => {
+                setExperiment(onDuplicateMetric(metric))
+                updateExperimentMetrics()
+            }}
         />
     )
