+++ frontend/src/scenes/experiments/experimentLogic.tsx
@@ -72,7 +72,6 @@ import { EXPERIMENT_MIN_EXPOSURES_FOR_RESULTS, MetricInsightId } from './constan
 import type { experimentLogicType } from './experimentLogicType'
 import { experimentsLogic } from './experimentsLogic'
 import { holdoutsLogic } from './holdoutsLogic'
-import { getDefaultMetricTitle } from './MetricsView/utils'
 import { SharedMetric } from './SharedMetrics/sharedMetricLogic'
 import { sharedMetricsLogic } from './SharedMetrics/sharedMetricsLogic'
 import { featureFlagEligibleForExperiment, percentageDistribution, transformFiltersForWinningVariant } from './utils'
@@ -405,10 +404,6 @@ export const experimentLogic = kea<experimentLogicType>([
             sharedMetricIds,
             metadata,
         }),
-        duplicateMetric: (
-            metric: ExperimentMetric | ExperimentTrendsQuery | ExperimentFunnelsQuery,
-            isPrimary: boolean
-        ) => ({ metric, isPrimary }),
         removeSharedMetricFromExperiment: (sharedMetricId: SharedMetric['id']) => ({ sharedMetricId }),
         createExperimentDashboard: true,
         setIsCreatingExperimentDashboard: (isCreating: boolean) => ({ isCreating }),
@@ -426,24 +421,6 @@ export const experimentLogic = kea<experimentLogicType>([
             {
                 setExperiment: (state, { experiment }) => {
                     return { ...state, ...experiment }
-                },
-                duplicateMetric: (state, { metric, isPrimary }) => {
-                    const name = metric.name
-                        ? `${metric.name} (copy)`
-                        : metric.kind === NodeKind.ExperimentMetric
-                        ? `${getDefaultMetricTitle(metric)} (copy)`
-                        : undefined
-
-                    const newMetric = { ...metric, id: undefined, name }
-                    const metricsKey = isPrimary ? 'metrics' : 'metrics_secondary'
-                    const metrics = [...state[metricsKey]]
-                    const originalIndex = metrics.findIndex((m) => m === metric)
-                    metrics.splice(originalIndex + 1, 0, newMetric)
-
-                    return {
-                        ...state,
-                        [metricsKey]: metrics,
-                    }
                 },
                 addVariant: (state) => {
                     if (state?.parameters?.feature_flag_variants) {
@@ -848,13 +825,6 @@ export const experimentLogic = kea<experimentLogicType>([
         ],
     }),
     listeners(({ values, actions }) => ({
-        duplicateMetric: ({ isPrimary }) => {
-            if (isPrimary) {
-                actions.loadMetricResults()
-            } else {
-                actions.loadSecondaryMetricResults()
-            }
-        },
         createExperiment: async ({ draft, folder }) => {
             const { recommendedRunningTime, recommendedSampleSize, minimumDetectableEffect } = values
